delete groovy inactive message

+++ src/commands/move.js
@@ -9,19 +9,28 @@ module.exports = {
 
         // Id of MEE6 and Groovy
         if (message.author.id == '159985870458322944' || message.author.id == '234395307759108106') {
-            console.log(message.embeds);
-            channel.send(message.embeds)
-                .then(console.log(`Embed from other bots moved successfully`))
+            if (message.embeds.description.includes("https://groovy.bot/premium")) {
+                message.delete()
+                .then(console.log(`Groovy inactivity message deleted successfully`))
                 .catch((error) => {
                     throw error;
                 });
-        }
 
-        message.delete()
-            .then(console.log(`Message deleted successfully`))
+                return;
+            }
+            
+            channel.send(message.embeds)
+            .then(console.log(`Embed from other bots moved successfully`))
             .catch((error) => {
                 throw error;
             });
+        }
+
+        message.delete()
+        .then(console.log(`Message deleted successfully`))
+        .catch((error) => {
+            throw error;
+        });
 
         return;
     }
